requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
contourpy==1.1.0
cycler==0.11.0
filelock==3.9.0
fonttools==4.42.1
idna==3.4
imutils==0.5.4
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
MarkupSafe==2.1.2
matplotlib==3.7.2
mpmath==1.2.1
networkx==3.0
numpy==1.25.2
opencv-python==4.8.0.76
packaging==23.1
Pillow==9.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.28.1
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
sympy==1.11.1
threadpoolctl==3.2.0
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchvision==0.15.2+cu118
typing_extensions==4.4.0
urllib3==1.26.13
